feat: add trivy scan1

+++ Jenkinsfile
@@ -47,21 +47,24 @@ pipeline {
                 }
             }
         }
-        stage('Scan Trivy') {
+        stage('Security Scan (Trivy)') {
             steps {
                 script {
                     def isMR = (env.gitlabMergeRequestIid != null || env.CHANGE_ID != null)
                     def isMaster = (env.BRANCH_NAME == 'master' || env.BRANCH_NAME == 'main')
                     def isTag = (env.TAG_NAME != null)
-                    def buildCond = (isMR || isMaster || isTag)
 
-                    conditionalStage(name: 'Trivy SCA', condition: buildCond) {
-                        echo "Scanning image for vulnerabilities: ${IMAGE_NAME}"
+                    if (isMR || isMaster || isTag) {
+                        echo "Trigger detected! Running Trivy for ${IMAGE_NAME}..."
                         sh "docker run --rm -v /var/run/docker.sock:/var/run/docker.sock aquasec/trivy:latest image --format sarif --output trivy_report.sarif ${IMAGE_NAME}"
                         archiveArtifacts artifacts: 'trivy_report.sarif', allowEmptyArchive: true
-                        recordIssues(
-                            tools: [sarif(pattern: 'trivy_report.sarif', id: 'trivy', name: 'Trivy SCA Scan')]
-                        )
+                    recordIssues(
+                        tools: [sarif(pattern: 'trivy_report.sarif', id: 'trivy', name: 'Trivy SCA Scan')]
+                    )
+                    } else {
+                        Utils.markStageSkippedForConditional('Security Scan (Trivy)')
+                        echo "Skipping Trivy: Not a target branch or MR."
+                        return
                     }
                 }
             }
